+++ config/livewire-tables.php
@@ -2,7 +2,7 @@
 
 return [
     /**
-     * Options: tailwind | bootstrap-4.
+     * Options: tailwind | bootstrap-4 | bootstrap-5.
      */
     'theme' => 'tailwind',
 ];