+++ resources/views/bootstrap-5/components/table/row-columns.blade.php
@@ -1,0 +1,9 @@
+@foreach($columns as $column)
+    <td>
+        @if ($column->asHtml)
+            {{ new \Illuminate\Support\HtmlString($column->formatted($row)) }}
+        @else
+            {{ $column->formatted($row) }}
+        @endif
+    </td>
+@endforeach